Crate fitsio_sys

This package was automatically generated with rust-bindgen and as such was not user-generated.

The functions contained are expected to be used with fitsio, a high level API wrapper around the low level direct C-bindings, though the bindings are complete enough to be usable.

This code will not be directly documented, and so users should refer to the fitsio C documentation for usage.

Note about function names

Unfortunately we must use fits short names throughout. The C-api exposes long names for functions which are more descriptive, for example fits_open_file instead of ffopen, but the symbols available in the library have only short names, and the long names are merely preprocessor definitions.

Examples

use std::ptr;
use std::ffi;

let filename = ffi::CString::new("!/tmp/test.fits").unwrap();
let mut fptr = ptr::null_mut();
let mut status = 0;

unsafe {
    // Create a new file, clobbering any pre-existing file
    ffinit(&mut fptr as *mut *mut _,
        filename.as_ptr(),
        &mut status);

    // Add an empty primary HDU
    ffphps(fptr, 8, 0, ptr::null_mut(), &mut status);

    // Finally close the file
    ffclos(fptr, &mut status);
}

assert_eq!(status, 0);

Structs

FITSfile

PixelFilter

fitsfile

iteratorCol

tcolumn

wtbarr

Enums

CaseSensitivity

DataType

FileOpenMode

HduType

ImageType

Statics

MAX_VALUE_LENGTH

Functions

CFITS2Unit

CUnit2FITS

ffasfm

ffbinr

ffbins

ffbnfm

ffbnfmll

ffcalc

ffcalc_rng

ffcdfl

ffchfl

ffclos

ffcmph

ffcmps

ffcmrk

ffcmsg

ffcnvthdr2str

ffcopy

ffcpcl

ffcpdt

ffcpfl

ffcphd

ffcpky

ffcprw

ffcrhd

ffcrim

ffcrimll

ffcrow

ffcrtb

ffcsum

ffdcol

ffdelt

ffdhdu

ffdkey

ffdkinit

ffdkopn

ffdopn

ffdrec

ffdrow

ffdrrg

ffdrws

ffdrwsll

ffdstr

ffdsum

ffdt2s

ffdtdm

ffdtdmll

ffdtyp

ffeqty

ffeqtyll

ffesum

ffexist

ffextn

ffexts

ffffrw

fffkls

ffflmd

ffflnm

ffflsh

ffflus

fffree

fffrow

fffrwc

ffg2db

ffg2dd

ffg2de

ffg2di

ffg2dj

ffg2djj

ffg2dk

ffg2dsb

ffg2dui

ffg2duj

ffg2duk

ffg3db

ffg3dd

ffg3de

ffg3di

ffg3dj

ffg3djj

ffg3dk

ffg3dsb

ffg3dui

ffg3duj

ffg3duk

ffgabc

ffgacl

ffgbcl

ffgbclll

ffgcdw

ffgcf

ffgcfb

ffgcfc

ffgcfd

ffgcfe

ffgcfi

ffgcfj

ffgcfjj

ffgcfk

ffgcfl

ffgcfm

ffgcfs

ffgcfsb

ffgcfui

ffgcfuj

ffgcfuk

ffgcks

ffgcl

ffgcnn

ffgcno

ffgcrd

ffgcv

ffgcvb

ffgcvc

ffgcvd

ffgcve

ffgcvi

ffgcvj

ffgcvjj

ffgcvk

ffgcvl

ffgcvm

ffgcvs

ffgcvsb

ffgcvui

ffgcvuj

ffgcvuk

ffgcx

ffgcxui

ffgcxuk

ffgdes

ffgdesll

ffgdess

ffgdessll

ffgerr

ffgextn

ffggpb

ffggpd

ffggpe

ffggpi

ffggpj

ffggpjj

ffggpk

ffggpsb

ffggpui

ffggpuj

ffggpuk

ffghad

ffghadll

ffghbn

ffghbnll

ffghdn

ffghdt

ffghof

ffghpr

ffghprll

ffghps

ffghsp

ffghtb

ffghtbll

ffgics

ffgicsa

ffgidm

ffgidt

ffgiet

ffgipr

ffgiprll

ffgisz

ffgiszll

ffgiwcs

ffgkcl

ffgkey

ffgkls

ffgknd

ffgkne

ffgknj

ffgknjj

ffgknl

ffgknm

ffgkns

ffgky

ffgkyc

ffgkyd

ffgkye

ffgkyj

ffgkyjj

ffgkyl

ffgkym

ffgkyn

ffgkys

ffgkyt

ffgmcp

ffgmng

ffgmop

ffgmrm

ffgmsg

ffgmtf

ffgncl

ffgnrw

ffgnrwll

ffgnxk

ffgpf

ffgpfb

ffgpfd

ffgpfe

ffgpfi

ffgpfj

ffgpfjj

ffgpfk

ffgpfsb

ffgpfui

ffgpfuj

ffgpfuk

ffgpv

ffgpvb

ffgpvd

ffgpve

ffgpvi

ffgpvj

ffgpvjj

ffgpvk

ffgpvsb

ffgpvui

ffgpvuj

ffgpvuk

ffgpxf

ffgpxfll

ffgpxv

ffgpxvll

ffgrec

ffgrsz

ffgsdt

ffgsfb

ffgsfd

ffgsfe

ffgsfi

ffgsfj

ffgsfjj

ffgsfk

ffgsfsb

ffgsfui

ffgsfuj

ffgsfuk

ffgstm

ffgstr

ffgsv

ffgsvb

ffgsvd

ffgsve

ffgsvi

ffgsvj

ffgsvjj

ffgsvk

ffgsvsb

ffgsvui

ffgsvuj

ffgsvuk

ffgtam

ffgtbb

ffgtch

ffgtcl

ffgtclll

ffgtcm

ffgtcp

ffgtcr

ffgtcs

ffgtdm

ffgtdmll

ffgthd

ffgtis

ffgtmg

ffgtnm

ffgtop

ffgtrm

ffgtvf

ffgtwcs

ffgunt

ffh2st

ffhdef

ffhdr2str

ffhist

ffhist2

ffibin

fficls

fficol

ffifile

ffifile2

ffiimg

ffiimgll

ffikey

ffikfc

ffikfm

ffikls

ffikyc

ffikyd

ffikye

ffikyf

ffikyg

ffikyj

ffikyl

ffikym

ffikys

ffikyu

ffimem

ffimport_file

ffinit

ffinttyp

ffiopn

ffirec

ffirow

ffitab

ffiter

ffiurl

ffkeyn

ffmahd

ffmaky

ffmbyt

ffmcom

ffmcrd

ffmkfc

ffmkfm

ffmkky

ffmkls

ffmkyc

ffmkyd

ffmkye

ffmkyf

ffmkyg

ffmkyj

ffmkyl

ffmkym

ffmkys

ffmkyu

ffmnam

ffmnhd

ffmrec

ffmrhd

ffmrky

ffmvec

ffnchk

ffnkey

ffomem

ffopen

ffopentest

ffp2db

ffp2dd

ffp2de

ffp2di

ffp2dj

ffp2djj

ffp2dk

ffp2dsb

ffp2dui

ffp2duj

ffp2duk

ffp3db

ffp3dd

ffp3de

ffp3di

ffp3dj

ffp3djj

ffp3dk

ffp3dsb

ffp3dui

ffp3duj

ffp3duk

ffpcks

ffpcl

ffpclb

ffpclc

ffpcld

ffpcle

ffpcli

ffpclj

ffpcljj

ffpclk

ffpcll

ffpclm

ffpcls

ffpclsb

ffpclu

ffpclui

ffpcluj

ffpcluk

ffpclx

ffpcn

ffpcnb

ffpcnd

ffpcne

ffpcni

ffpcnj

ffpcnjj

ffpcnk

ffpcnl

ffpcns

ffpcnsb

ffpcnui

ffpcnuj

ffpcnuk

ffpcom

ffpdat

ffpdes

ffpextn

ffpgpb

ffpgpd

ffpgpe

ffpgpi

ffpgpj

ffpgpjj

ffpgpk

ffpgpsb

ffpgpui

ffpgpuj

ffpgpuk

ffphbn

ffphext

ffphis

ffphpr

ffphprll

ffphps

ffphpsll

ffphtb

ffpkfc

ffpkfm

ffpkls

ffpknd

ffpkne

ffpknf

ffpkng

ffpknj

ffpknjj

ffpknl

ffpkns

ffpktp

ffpky

ffpkyc

ffpkyd

ffpkye

ffpkyf

ffpkyg

ffpkyj

ffpkyl

ffpkym

ffpkys

ffpkyt

ffpkyu

ffplsw

ffpmrk

ffpmsg

ffpnul

ffppn

ffppnb

ffppnd

ffppne

ffppni

ffppnj

ffppnjj

ffppnk

ffppnsb

ffppnui

ffppnuj

ffppnuk

ffppr

ffpprb

ffpprd

ffppre

ffppri

ffpprj

ffpprjj

ffpprk

ffpprn

ffpprsb

ffppru

ffpprui

ffppruj

ffppruk

ffppx

ffppxll

ffppxn

ffppxnll

ffprec

ffprwu

ffpscl

ffpss

ffpssb

ffpssd

ffpsse

ffpssi

ffpssj

ffpssjj

ffpssk

ffpsssb

ffpssui

ffpssuj

ffpssuk

ffpsvc

ffptbb

ffptdm

ffptdmll

ffpthp

ffpunt

ffrdef

ffreopen

ffrprt

ffrsim

ffrsimll

ffrtnm

ffrwrg

ffrwrgll

ffs2dt

ffs2tm

ffsnul

ffsrow

fftexp

ffthdu

fftheap

fftkey

fftm2s

fftnul

fftopn

fftplt

fftrec

fftscl

ffucrd

ffukfc

ffukfm

ffukls

ffuky

ffukyc

ffukyd

ffukye

ffukyf

ffukyg

ffukyj

ffukyl

ffukym

ffukys

ffukyu

ffupch

ffupck

ffurlt

ffvcks

ffverifydate

ffvers

ffwldp

ffwrhdu

ffxypx

fits_calc_binning

fits_comp_img

fits_compress_table

fits_copy_cell2image

fits_copy_image2cell

fits_copy_image_section

fits_copy_pixlist2image

fits_decomp_img

fits_delete_iraf_file

fits_execute_template

fits_get_compression_type

fits_get_dither_seed

fits_get_hcomp_scale

fits_get_hcomp_smooth

fits_get_noise_bits

fits_get_quantize_level

fits_get_section_range

fits_get_tile_dim

fits_get_token

fits_get_token2

fits_hcompress

fits_hcompress64

fits_hdecompress

fits_hdecompress64

fits_img_compress

fits_img_decompress

fits_img_decompress_header

fits_img_stats_float

fits_img_stats_int

fits_img_stats_short

fits_init_cfitsio

fits_is_compressed_image

fits_is_reentrant

fits_iter_get_array

fits_iter_get_colname

fits_iter_get_colnum

fits_iter_get_datatype

fits_iter_get_file

fits_iter_get_iotype

fits_iter_get_repeat

fits_iter_get_tdisp

fits_iter_get_tlmax

fits_iter_get_tlmin

fits_iter_get_tunit

fits_iter_set_by_name

fits_iter_set_by_num

fits_iter_set_colname

fits_iter_set_colnum

fits_iter_set_datatype

fits_iter_set_file

fits_iter_set_iotype

fits_make_hist

fits_pixel_filter

fits_read_wcstab

fits_rebin_wcs

fits_select_image_section

fits_set_compression_pref

fits_set_compression_type

fits_set_dither_offset

fits_set_dither_seed

fits_set_hcomp_scale

fits_set_hcomp_smooth

fits_set_huge_hdu

fits_set_lossy_int

fits_set_noise_bits

fits_set_quantize_dither

fits_set_quantize_level

fits_set_quantize_method

fits_set_tile_dim

fits_split_names

fits_translate_keyword

fits_translate_keywords

fits_uncompress_table

fits_write_keys_histo

Type Definitions

LONGLONG